Move TextUi to a separate package (to allow for other UIs in the future)
